ai-ceo/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
schedule==1.2.0
